requirements.txt
torch
transformers
python-dotenv
accelerate